improve firebase login
